mod/ipa_generation/requirements.txt
# RunPod serverless worker
runpod>=1.0.0

# Core dependencies for acoustic model inference
# GPU support via CUDA (install appropriate PyTorch version for your CUDA version)
# For RunPod: PyTorch with CUDA support is typically pre-installed
torch>=2.0.0

# Audio processing (optional, for audio-guided G2P)
librosa==0.9.2
soundfile
numpy

# ESPnet / POWSM (G2P)
espnet
espnet_model_zoo==0.1.7

# HTTP client for downloading audio files (if audio_uri provided)
requests>=2.31.0
